refactor(OmniView): Add shared datamodel for post page
